Add what I have
This is a prototype of a source-to-source translator that extract
functions from a C file with all its compilation dependencies. Or
it is what it is supposed to be.

Currectly, function tracking works correct, but TagTypes (structs,
enums, unions) seems to be carrying unecessary content. Global
variable tracking and typedefs are not working yet.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>