Fix array declaration of size as EnumConstantDecl
Handle the corner case where an array was declared as something like

enum {
  const1 = 1,
};

int var[const1];

in which clang will unfortunatelly expands const1 to 1 and lose the
EnumConstantDecl in this case, forcing us to reparse this declaration.

We fix this problem by implementing a special EnumTable which maps the
constant symbols to its respective EnumDecl, and then we mark this
EnumDecl for output.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>