Remove callgraph from Closure analysis.
Previously, we used a callgraph to find the set of functions we would
need to extract from the sourcecode. The problem is that some functions
are not called in the code, only referenced, thus never appearing in the
callgraph. See closure-1.c and closure-2.c testcases for an
illustration of what was happening.

To fix this, we search for function references in the body of the
functions. If we find those, we proceed with the analysis recursively,
removing the necessity of a callgraph.

This makes clang-extract successfuly extract dlsym from glibc.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>