Add partial support for C++ templates and namespace
clang-extract was designed with C in mind, although clang is able to
parse C++ as well. This is our first attempt to add support for C++.
However, because of the very recursive nature of C++ (namespaces inside
namespaces, typedefs inside classes and so on), supporting it will
require quite some refactoring work.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>