New `inline` tool
When livepatching, we need to find which symbols are available in the
target binary and which symbols are inlined.  This commit adds a new
tool called `inline` which is able to:

- Show all available symbols in the target binary.
- Show the symbols which got inlined.
- Show all symbols got inlined into a function.
- Show where a symbol is inlined.

For this analysis the tool uses the `.ipa-clones` files together with
the debuginfo files. The tools supports -flto by analysing multiple
.ipa-clones altough it doesn't handle some quirks of LTO
transformations.