Add support for typedef to a template specifier
In C++ you can typedef to a tmeplate specifier. For example:

template<typename _CharT>
  struct _Char_types
  {
    typedef unsigned long int_type;
  };

template<typename _CharT>
struct char_traits
{
  typedef typename _Char_types<_CharT>::int_type int_type;
};

Here is clear that if we want to add `struct char_traits`, we must also
drag `struct _Char_types` with it.  In C, the method
getAnonDeclWithTypedefName() returned the name of the Record in question,
but here is simply returns NULL.  In order to correctly add those we
need to use the information encoded in an ElaboratedType, which was not
trivial to find at first.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>