Fix incomplete definition of structs in kernel
clang's getPrevious seems broken. It is supposed to get the parent node
of another node in the AST, but When trying to add `struct key;` from kernel
6.6-rc7 defined in `include/linux/key.h:33`, getPrevious returns
`task_struct` which is nonsense: it is not by any means a child of
struct key.  Hence we need to open a bug in clang about it (but first
check if it was fixed in llvm-17 or llvm-18).  See issue #12.

This is a regression on C++ support, which is broken ATM anyway.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>