Fix SymbolExternalizer missing references to symbol when -DCE_RENAME_SYMBOLS in on.
Previously, on FunctionUpdater, we only checked of the reference
matched the name of the Declaration. But when we are just renaming the
symbol, we only swaps its identifier, not creating a new node for the
AST. Hence we have to actually check in the Source Text what the user
wrote there and update it as well.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>