InlineAnalisys: Kernel: Externalize symbols that are not from vmlinux
When creating a LP for kernel we should externalize everything that
does not come from vmlinux, since everything else is a symbol that is
either private, or is public but comes from a different module.

Depending on symbols from different modules incurs that our LP would
load the module where the function comes from before loading the LP,
which is not what we want.

A server would not care if we patch a function from the bluetooth
module, because could not own any bluetooth devices, so loading the
bluetooth module may incur in a new attack surface would the need of it
being loaded.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>