Expand '~' characters in path.
clang-extract uses -DCE_ARG=<PATH> in several cases, which the shell do
not expand.  This commit expand the '~' characters in paths by adding
code in the "Parser" constructor to do it.

It also fixes a memory leak in the case where the InlineAnalysis object
wasn't fully constructed (exception occured when constructing it).

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>