Avoid weak externalization of functions without a prototype
It seems that clang (at least 16) have a bug in the Rewriter.

on include/linux/fortify-string.h from linux-5.14.21-150500.55.28

With the correct SourceRange pointing to from line 179 to 192, it
incorrectly erases lines 163 to 192. This seems to be a bug in the
rewriter and a bug report to clang should be open with a reproducer.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>