Add meson build files
To work with meson, execute

$ meson setup build --native-file ce-native.ini

Which will run the config phase, using clang++ as the compiler. To
compile, just execute

$ meson compile -C build

Now the check target from Makefile expects the binary to be created
inside the build dir, as stated in the setup phase above.

For now it detects if g++ is being used, and fails the setup phase if
clang++ wasn't used.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>