Introduce a conflict resolver in SymbolExternalizer
Clang's Rewriter assumes that all changes are commutative, that means
A + B = B + A. This is not the case when whe are externalizing things,
once we may externalize a function that in its body contains a reference
to another externalized function. Hence we have to cleverly find which
modifications stays and which should be discarded.

We do this by setting a priority in each Text Modifications, and if two
modifications wants to modify the same piece of text, we only let the one
with highest priority.  With this, we classify the modifications which
removes content to be the highest priority, and the ones which only changes
the name of things to be the lowest priority.  Hence at some point we compute
the insersection of all with all and discards the ones with lower priority.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>