Fix Closure and Externalization error on vmwgfx_surface.c
When attempting to analyze vmwgfx_surface.c from kernel 5.14, there was
a closure computation error and a Symbol Externalization error.  The
first because somehow the RecursiveASTVisitor did not parse the
definition of a static function we marked for output, and the second
because we forgot to include '\n' as a space token.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>