Fix clang discarding changes to a modified file
The Rewriter class logs changes to the FileID, and not the FileEntry.
A single FileEntry can have multiple FileIDs, and
SourceManager::translateFile() is able to get a FileID from a FileEntry.

But it turns out that this FileID may not be the one that contains the
changes we must apply to the file, hence we now avoid using translateFile
(poison the symbol) and use our own map instead (FileEntryMap) which
we maps the FileEntry to a suitable FileID.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>