Fix header expansion including headers not reachable from the main file
When marking certain includes to be output, we must be careful if the
include is reachable from the main file, otherwise the generated file
won't be able to compile with the same flags passed to the original
file.  We fix this by checking if the header file in question is
reachable from the main file when it is set to output, and if not we
mark it for expansion.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>