Add boilerplate code for IncludeExpansionPolicy
This commit adds a series of classes which can be used to decide if a
given #include should be expanded or not based on project policy.

Now passing -DCE_EXPAND_INCLUDES=nothing, no additional includes will
be expanded. while as -DCE_EXPAND_INCLUDES=kernel will use the kernel
policy.

Note that passing -DCE_EXPAND_INCLUDES alone still works and will
either set it to nothing or kernel depending if -D__KERNEL__ was
passed or not.

-DCE_EXPAND_INCLUDES=everything is only used for testing purposes as it
will trigger the expansion of everything, which is already supported by
not passing -DCE_EXPAND_INCLUDES.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>