Fix FunctionDepsFinder only analyzing the last Decl of a requested symbol
Previously, FunctionDepsFinder::Find_Functions_Required sometimes only
analyzed the last declaration of a function, which breaks in the case
where the function has a new prototype defined later of its body.  The
reason for this is that DeclContext::lookup fails to find all symbols
of a given name.

If this is a result of a bug in LLVM this is yet to be determined.

Fixes #20.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>