Add script to parse klp-build's ce.out.txt log and generate a runscript from it
klp-build is a series of scripts to generate kernel livepatches. It
supports running clang-extract as a backend for extracting code from
the kernel.  However, when things break inside clang-extract we must
extract the interesting running commands from its logs and cd into
the build dir.  This script facilitates that by generating a .sh
script which can be used to run clang-extract easily.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>