Fix externalization in command line macros, IBT tests
This commit fixes a weird case where the SymbolExternalizer attempts
to rewrite a macro that was defined in the command line.

This also fixes the first IBT testcase, as well as add new ones that
are currently failing so we mark them as XFAIL.

linux tests are now in testsuite/linux.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>