Fix dropped attributes from vector types
Clang handle __attribute__((__vector_size__((N)))) attributes differently
from the general case.  While other attributes resides in the AttrVec
structure, those are hardcoded in the VectorType structure and the AttrVec
may contain nothing.  Hence we also need to trigger the Get_Expanded_Loc
for this case as well even if we can't get the SourceLocation of this
attribute.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>