Fix nested struct being discarded if parent struct is flagged as partial
When clang detects that a certain struct doesn't need a full definition,
i.e. its size is not required, then the flag CompleteDefinitionRequired
is set to false.  But in the following case:
```
struct Fts3Table {
  struct Fts3Index {
    int nPrefix;
  } *aIndex;
};

int fts3PrefixParameter(
  const char *zParam,
  int *pnIndex,
  struct Fts3Index **apIndex
){
  int a = sizeof(struct Fts3Index);
  return 0;
}
```
this results in the nested struct `Fts3Index` not being output as its
parent struct `Fts3Table` is marked as CompleteDefinitionRequired as
false.  In such cases we have to force CompleteDefinitionRequired of
the parent struct as true so that its inner structs can be output.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>