Avoid adding previous declarations of typedefs into closure
This bizarre kind of constructs:
```
typedef __builtin_va_list __gnuc_va_list; // 1
typedef __gnuc_va_list va_list;           // 2
typedef __builtin_va_list va_list;        // 3
```
Is accepted by both clang and gcc. If a function references va_list,
the last decl will reference (3), but since we always add the previous
decl version to the closure as well we end up adding (2) but not (1).
Looking closely we figure out that (2) is not necessary, thus we avoid
adding the previous version of the decls to the closure.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>