Deference pointer types when searching for decomposed typedef-struct
clang decomposes:
```
typedef struct _fakestr {
  struct _fakestr *pNext;
  int type;
} FAKE, *PFAKE;
```
into the AST equivalent of
```
struct _fakestr {
  struct _fakestr *pNext;
  int type;
};

typedef struct _fakestr FAKE, *PFAKE;
```
but since we output what the user wrote, we must remove the
`struct _fakestr` Decl from the closure.  But in this case
`PFAKE` is a pointer to `_fakestr` so we need to deference
it to get the definition of `_fakestr`.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>