Add (failing) testcase
small/enum-10.c fails because it declares two global variables:
```
int a[NB_END], b[INB_END];
```
but only references `b`, hence `a` is discarded by closure but
PrettyPrint will print it nonetheless because it is necessary to reach
`b`. But since `a` was discarded, the enum in which `NB_END` is declared
is discarded as well.

This can be workaround by passing `-DCE_EXPORT_FUNCTIONS=f,a` so it will
be forced into the closure.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>