libcextract: Add support for Weak symbols for kernel mode
Until this point all symbols found in kernel mode (usually using Symvers
file), were strong (with externalized variables) or none (which copied
the symbols to the closure). Now clang-extract checks is the symbol is
present on vmlinux, meaning that the symbol is weak externalized,
meaning the the symbol will be always present.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>