Disable Precompiled Headers
Clang provides a feature to speedup header parsing which consists of
generating a PrecompiledHeader (.pch) file to avoid having to compile
the header multiple times if the main file needs to be reparsed
multiple times, which is the case we are interested.

However, there seems to be a bug there because the Preprocessor is
re-instantiated after generating them, loosing the __COUNT__ counter
value. So we disable it until this issue is fixed.

After testing PCH enabled vs. disabled, I noticed clang-extract is
significantly *faster* if PCH is disabled, hence disable it once
for all (1.6s vs. 1.0s for vmwgfx_execbuf case).

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>