Fix CompleteDefinition of struct being dragged into output because of union field
On the following example:
```
struct ll {
  union {
    unsigned long key;
    void *key_ptr;
  };
  struct ll *next;
};
```
Even if we don't need the full definition of `ll` it is dragged into
the output because when analyzing this `union` field it tries to go
up and set `struct ll` as having its body necessary once there was
no way for it to know that it went there as a consequence of adding
`struct ll` to the closure, rather than it being there because of
a foward union declaration inside a struct.  This commit fixes
this by pushing the Decls into a stack as a way to know which
declarations we are currently analyzing.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>