Fix Late Externalization when variable is declared as macro expansion
In the case where a variable is declared as a consequence of a macro
expansion, as an example:
```
 #include "lateext-7.h"
 DEFINE_STATIC_KEY_FALSE(aaa);
```
What happens is that when trying to get the FileID of `aaa` it returns
the FileID for the include file rather than the main file.  Hence make
sure it gets the expansion location rather than the spelling location.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>