libcextract: ExpansionPolicy: Check for abs and rel path differences
There are two situations when extract kernel code: using SLE codestreams
or extracting from a built kernel source directory. For the previous
one, for some reason, absolute_path and relative_path are the same, so
we need to check if "/include/" path is present in the include path is
there or not to expand the header.

For the kernel source tree things are different: they are different, so
we can check if the relative path is "./include", and thus things are
decided correctly with the fix.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>