Disable const propagation on externalized vars
The value of the externalized vars are filled by the livepatch system,
not the compiler.  Hence if there are no writes into the variable
the compiler won't generate the instructions to read it.  As an
example:

```
static int aa;
int f(void) { return aa; }
```
generates:
```
  xor eax, eax
  ret
```
Notice how there is no memory read.  Adding the `__attribute__((used))`
to it changes the asm code to:
```
  movl	aa(%rip), %eax
  ret
```
Fixing the issue.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>