Add option to pass two debuginfos
SUSE packages contain the stripped .so file with .dynsym and the
debuginfo (.debug) provided in -debuginfo packages.  We need both
of them in Userspace Livepatching since 15.5 for proper analysis.
This commit patches the ElfCache and InlineAnalysis so it can use
both of them.

For kernel we expect it to only use one ELF file for now (the
first one).

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>