Analyze partial template declarations
C++ can specify partial template declarations which can implement
additional things that the original declaration doesn't have.  Here
we try to match if a certain declaration comes from a specific
specification by matching the Identifier on a DependentNameType, but
that is not always possible so we also add all
ClassTemplatePartialSpecializationDecl for a ClassTemplateDecl to the
closure until we do not find a better way of doing this.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>