Fix Remove_Redundant_Decls removing necessary decls
PrettyPrint::Contains_From_LineCol had a bug in which Decls that do not
contain themselves were marked as such because two files ended up having
the same FileID because of preprocessed notation (example: # 1 "file").
Fix this by dropping this function in favor of
Range_Fully_Contains_Range, which uses the expanded location.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>