Make sure we get a matching number of #if and #endifs before copying function to output
glibc has some weird cornercases in `sysdeps/generic/ldsodefs.h` where
a struct definition depends of a macro definition (SHARED), but its
fields attributes does not.  Here we devise an heuristic to try to
detect such cases by looking if the number of #ifs and #endifs matches
in the function.  If not, then fallback to AST dumping.

Closes #147

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>