Remove any use of `basename`
The function `basename` has 2 versions, one strdup the input string and
the other one just returns a pointer to the original string.  The one
which is being used is not clear in the program and we were using both
of them in this project, hence get rid of it entirely and provide our
own.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>