Add support to avoid header expansion
Previously, clang-extract only properly supported flags to expand
headers. Now the new flag:
```
-DCE_NOT_EXPAND_INCLUDES=header1.h,...,headern.h
```
can be used to instruct it to not expand certain headers.  To enable it,
the expansion must go through the -DCE_KEEP_INCLUDES engine, which
means it is only activated if a certain policy is enabled.

Furthermore, this commit also adds a new program: `ce-includetree`
which shows in stdout the IncludeTree of the input source file.
it can be invoked using the compilation flags of the original file.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>